requirements.txt
locustio==0.8.1
pyzmq==17.0.0